car is working

+++ worldclass/src/main/java/com/worldclass/listeners/BallPosListener.java
@@ -4,5 +4,5 @@ package com.worldclass.listeners;
  * Created by erz on 4/6/2014.
  */
 public interface BallPosListener {
-    public int getBallX();
+    public float getBallX();
 }
